refactored code. moved utility functions from db-manager folder to root repo folder. Added function to fetch ICX balance of a wallet
