+++ package.json
@@ -25,8 +25,10 @@
   },
   "dependencies": {
     "dotenv": "^16.4.5",
+    "icon-sdk-js": "^1.5.2",
     "mocha": "^10.4.0",
     "mongodb-memory-server": "^9.1.8",
-    "mongoose": "^8.3.2"
+    "mongoose": "^8.3.2",
+    "rqst": "github:FidelVe/rqst"
   }
 }
