+++ utils/json-rpc-services.js
@@ -1,5 +1,5 @@
 const { makeJsonRpcRequestObject, makeJsonRpcCall } = require("./utils");
-const config = require("../../utils/config");
+const config = require("./config");
 
 async function getNetworkInfo(height = null) {
   try {
@@ -201,9 +201,21 @@ async function getUserRegistrationBlock(
   }
 }
 
-async function getIcxBalance(wallet) {
-  void wallet;
-  return 0;
+async function getIcxBalance(wallet, height = null) {
+  try {
+    const requestObj = makeJsonRpcRequestObject(
+      null,
+      null,
+      null,
+      height,
+      "icx_getBalance",
+      { address: wallet },
+    );
+    return await makeJsonRpcCall(requestObj, config.jvm.default.rpc);
+  } catch (e) {
+    console.log("Error making icx_getBalance request");
+    console.error(e);
+  }
 }
 
 module.exports = {
